refactor: ChatDetailLayout 추가
