+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/ChatDetailLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ChatDetailLayout';
